carry contextual information with the DIECursor
DWARF5 has more contextual information that is associated with the
compilation unit. As a preparation for using such information, carry it
with the DIECursor and eliminate places where we're passing in the
parent compilation unit.

Also add the RDAddr helper to read a target-address according to the
specification in the compilation unit.